Move CuratorMutex to make it reusable
